% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCVT.R
\name{drawCVT}
\alias{drawCVT}
\title{Draws a WFP constant-volatility transformation}
\usage{
drawCVT(R, n = dim(R)[2], m = dim(R)[1])
}
\arguments{
\item{R}{Compositional matrix. See \code{\link{CVTest}}}

\item{n}{number of species}

\item{m}{number of timepoints}
}
\value{
two member list, "f" - the CVT evaluated at all points in the input matrix - and "DF", the jumps in the CVT
}
\description{
Draws a WFP constant-volatility transformation
}

