% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbmMR.R
\name{gbmMR}
\alias{gbmMR}
\title{Mean-Reverting geometric brownian motion}
\usage{
gbmMR(nspecies, mu = 1, sigma = 1, rho = 0, Tmax = 1, X0 = NULL,
  nsamples = 1000, dt = NULL)
}
\arguments{
\item{nspecies}{number of companies to simulate}

\item{mu}{Number or vector of length \code{nspecies} for the drift}

\item{sigma}{Number of vector of length \code{nspecies} for the volatility}

\item{Tmax}{End of time-window for simulation}

\item{X0}{initial abundances}

\item{nsamples}{Number of samples to collect between [0,Tmax]}

\item{dt}{size of timesteps - an alternative to nsamples for specifying numerical integration. If input, the output will be of length Tmax/dt}
}
\description{
Mean-Reverting geometric brownian motion
}
\examples{
library(WrightFisher)
library(plotrix)

x <- gbmMR(10)
par(mfrow=c(1,2))
stackpoly(x$Market,stack=T,main='Market Dynamics')
stackpoly(x$Shares,stack=T,main='Market Share Dynamics')
}

