% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WashburneNTest.R
\name{WashburneNTest}
\alias{WashburneNTest}
\title{Performs Washburne et al.'s (2016) Neutrality Test with corrected ks-test}
\usage{
WashburneNTest(R, ntests = min(16000, 2^dim(X)[2]), regress = "f",
  ncores = NULL, formula = DF ~ f + I(f^2), varformula = ~f + I(f^2), ...)
}
\description{
Performs Washburne et al.'s (2016) Neutrality Test with corrected ks-test
}

