% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfp.R
\name{wfp}
\alias{wfp}
\title{Simulate Wright-Fisher Process}
\usage{
wfp(nspecies = 3, rho = NULL, lambda = 15, X0 = rho, dt = 1e-05,
  Tmax = 1, nsamples = 1000, method = "VSM", tol = 1e-16)
}
\arguments{
\item{nspecies}{One way to parameterize WFP by number of species. \code{rho} will be \code{rep(1/nspecies,nspecies)}}

\item{rho}{Compositional vector signifying the mean of the community}

\item{lambda}{Rate of migration or mean-reversion}

\item{X0}{optional initial community state - must be a compositional vector of same size as rho}

\item{dt}{size of timesteps for numerical integration}

\item{Tmax}{Size of time window of numerical integration}

\item{nsamples}{Number of timepoints between 0 and T to sample. Output will be a matrix with nrow=length(rho) and ncol=nsamples}

\item{method}{Method for simulation, either 'VSM' or 'WFP'. Default is VSM - simulates volatility stabilized market and converts to relative abundances.}

\item{tol}{Tolerance for VSM simulation, i.e. the reflecting lower bound for stock prices to ensure values stay positive.}
}
\value{
Outputs a two-element list. The first element is the time vector for the time of samples, and the second element is the matrix of the fluctuating community.
}
\description{
Simulate Wright-Fisher Process
}
\examples{
library(plotrix)
X <- wfp(rho=rep(.2,5),dt=1e-6,Tmax=0.25)

time <- X$time
X <- X$Community
stackpoly(X,stack=T)
}
\references{
Washburne, Alex (2015) "Competition and Coexistence in an Unpredictable World". https://www.academia.edu/15517160/PhD_Thesis_-_Competition_and_Coexistence_in_an_Unpredictable_World
}

