% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WFP.R
\name{WFP}
\alias{WFP}
\title{Simulate Wright-Fisher Process}
\usage{
WFP(nSpecies = NULL, rho = rep(1/nSpecies, nSpecies), lambda = 15,
  X0 = rho, dt = 1e-05, Tmax = 1, sample = 1000)
}
\arguments{
\item{nSpecies}{One way to parameterize WFP by number of species. \code{rho} will be rep(1/nSpecies,nSpecies)}

\item{rho}{Compositional vector signifying the mean of the community}

\item{lambda}{Rate of migration or mean-reversion}

\item{X0}{optional initial community state - must be a compositional vector of same size as rho}

\item{dt}{size of timesteps for numerical integration}

\item{Tmax}{Size of time window of numerical integration}

\item{sample}{Number of timepoints between 0 and T to sample. Output will be a matrix with nrow=length(rho) and ncol=sample}
}
\value{
Outputs a two-element list. The first element is the time vector for the time of samples, and the second element is the matrix of the fluctuating community.
}
\description{
Simulate Wright-Fisher Process
}
\examples{
library(plotrix)
X <- WFP(rep(.2,5),dt=1e-6,T=0.25)

time <- X[[1]]
X <- X[[2]]
stackpoly(X,stack=T)
}
\references{
https://www.academia.edu/15517160/PhD_Thesis_-_Competition_and_Coexistence_in_an_Unpredictable_World
}

